Authorization Added

+++ routers/posts.py
@@ -6,21 +6,16 @@ from sqlalchemy.orm import selectinload
 import models
 from database import get_db
 from schemas import PostResponse, PostCreate, PostUpdate
+from auth import CurrentUser
 
 router = APIRouter()
 
 #Create Posts
 @router.post("", response_model=PostResponse, status_code=status.HTTP_201_CREATED)
 
-async def create_post(post: PostCreate, db : Annotated[AsyncSession, Depends(get_db)]):
-
-    result = await db.execute(select(models.User).where(models.User.id == post.user_id))
-    user = result.scalars().first()
-
-    if not user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User Not Found")
+async def create_post(post: PostCreate, current_user : CurrentUser, db : Annotated[AsyncSession, Depends(get_db)]):
     
-    new_post = models.Post(title=post.title, content=post.content, user_id=post.user_id)
+    new_post = models.Post(title=post.title, content=post.content, user_id=current_user.id)
 
     db.add(new_post) # Does not interact with database, no I/O -> does not need await
     await db.commit()
@@ -62,25 +57,19 @@ async def get_post(post_id : int, db : Annotated[AsyncSession, Depends(get_db)])
 # Update Post Fully
 @router.put("/{post_id}", response_model=PostResponse)
 
-async def update_post_fully(post_id : int, post_data : PostCreate, db : Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_fully(post_id : int, post_data : PostCreate, current_user : CurrentUser, db : Annotated[AsyncSession, Depends(get_db)]):
 
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
 
     if not post:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, details="Post Not Found")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post Not Found")
     
-    # Is the post being reassigned to another user? If yes → verify new user exists
-    if post_data.user_id != post.user_id:    
-        result = await db.execute(select(models.User).where(models.User.id == post_data.user_id))
-        user = result.scalars().first()
-
-        if not user:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, details="User Not Found")
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
         
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -91,13 +80,16 @@ async def update_post_fully(post_id : int, post_data : PostCreate, db : Annotate
 # Update post partially
 @router.patch("/{post_id}", response_model=PostResponse)
 
-async def update_post_partially(post_id : int, post_data : PostUpdate, db : Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partially(post_id : int, post_data : PostUpdate, current_user : CurrentUser, db : Annotated[AsyncSession, Depends(get_db)]):
 
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
 
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post Not Found")
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
     
     # Manual Update
 
@@ -127,13 +119,16 @@ async def update_post_partially(post_id : int, post_data : PostUpdate, db : Anno
 # Delete Post
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
 
-async def delete_post(post_id : int, db : Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id : int, current_user : CurrentUser, db : Annotated[AsyncSession, Depends(get_db)]):
 
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
 
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post Not Found")
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this post")
     
     await db.delete(post)
     await db.commit()   #Later on adding authentication we will allow only the author of the post to delete it.
